move cli logic to the cmd directory
